feat：Move `FeedLoaderPresentationAdapter` to separate file
